id INT to BIGINT

+++ ishtech-spring-boot-jwt-web/src/main/resources/db/migration/V20251021_020350__create_table_user_profile_aud.sql
@@ -1,6 +1,6 @@
 CREATE TABLE ${audit_schema_name}.t_user_profile_aud (
-  id            INT          NOT NULL,
-  rev           INT          NOT NULL,
+  id            BIGINT       NOT NULL,
+  rev           BIGINT       NOT NULL,
   revtype       SMALLINT     NOT NULL,
   rev_timestamp TIMESTAMP    NOT NULL DEFAULT CURRENT_TIMESTAMP,
   first_name    VARCHAR(255) NOT NULL,
